feat(time-picker): split core logic and bring pack timeChange event
