+++ packages/ng/time/ng-package.json
@@ -1,0 +1,7 @@
+{
+	"$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
+	"lib": {
+		"entryFile": "public-api.ts",
+		"styleIncludePaths": ["../styles"]
+	}
+}
