+++ packages/ng/time/public-api.ts
@@ -1,0 +1,1 @@
+export * from './time-picker/time-picker.component';
