+++ packages/ng/time/time-picker/time-picker.model.ts
@@ -1,6 +1,4 @@
-import { ISO8601Time } from './date-primitives';
-
-export type PickerControlDirection = 'up' | 'down';
+import { ISO8601Time } from '../core/date-primitives';
 
 export type TimeChangeEvent = {
 	previousValue: ISO8601Time | null;
@@ -12,9 +10,5 @@ export type TimeChangeEvent = {
 	| { source: 'control'; part: 'minutes' | 'hours'; direction: 'up' | 'down' }
 );
 export const DEFAULT_MIN_TIME: ISO8601Time = '00:00:00';
-export const DEFAULT_MAX_TIME: ISO8601Time = '23:59:59';
-export const DEFAULT_MAX_DURATION: ISO8601Time = '24:00:00';
-export const DEFAULT_PICKER_STEP = 1;
 
 export const DEFAULT_TIME_DECIMAL_PIPE_FORMAT = '2.0-0';
-export const DEFAULT_DURATION_HOUR_DECIMAL_PIPE_FORMAT = '1.0-0';
