+++ packages/ng/time/time-picker/time-picker.translate.ts
@@ -1,5 +1,5 @@
 import { InjectionToken } from '@angular/core';
-import { ILuTranslation } from '../core/translate';
+import { ILuTranslation } from '../../core/translate';
 
 export const LU_TIME_PICKER_TRANSLATIONS = new InjectionToken('LuTimePickerTranslations', {
 	factory: () => luTimePickerTranslations,
@@ -8,7 +8,6 @@ export const LU_TIME_PICKER_TRANSLATIONS = new InjectionToken('LuTimePickerTrans
 export type TimePickerTranslations = {
 	timePickerDescriptionAccessibility: string;
 	timePickerHours: string;
-	timePickerHourSeparator: string;
 	timePickerTimeSeparator: string;
 	timePickerMinutes: string;
 };
@@ -17,14 +16,12 @@ export const luTimePickerTranslations: ILuTranslation<TimePickerTranslations> = 
 	en: {
 		timePickerDescriptionAccessibility: '%%TODO%%',
 		timePickerHours: '%%HEURES%%',
-		timePickerHourSeparator: ':',
 		timePickerTimeSeparator: 'h',
 		timePickerMinutes: '%%MINUTES%%',
 	},
 	fr: {
 		timePickerDescriptionAccessibility: '%%TODO%%',
 		timePickerHours: '%%HEURES%%',
-		timePickerHourSeparator: ':',
 		timePickerTimeSeparator: 'h',
 		timePickerMinutes: '%%MINUTES%%',
 	},
